# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Your API",
    "description": "A simple example ASP.NET Core Web API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Societe": {
      "get": {
        "tags": [
          "Societe"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Societe"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Societe/{id}": {
      "get": {
        "tags": [
          "Societe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Societe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SocieteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Societe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Societe/name/{nomSociete}": {
      "get": {
        "tags": [
          "Societe"
        ],
        "parameters": [
          {
            "name": "nomSociete",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocieteDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Societe/adresse/{adresse}": {
      "get": {
        "tags": [
          "Societe"
        ],
        "parameters": [
          {
            "name": "adresse",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocieteDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Vente": {
      "get": {
        "tags": [
          "Vente"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vente"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Vente/{id}": {
      "get": {
        "tags": [
          "Vente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDetailsDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VenteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Vente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Vente/search": {
      "get": {
        "tags": [
          "Vente"
        ],
        "parameters": [
          {
            "name": "client",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "marque",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typeArticle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reference",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenteDetailsDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DetailsPaiementDTO": {
        "required": [
          "nCheque",
          "nlcn"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "venteId": {
            "type": "integer",
            "format": "int32"
          },
          "nCheque": {
            "type": "string",
            "nullable": true
          },
          "nlcn": {
            "type": "string",
            "nullable": true
          },
          "dateEcheance": {
            "type": "string",
            "format": "date-time"
          },
          "montant": {
            "type": "number",
            "format": "double"
          },
          "typeDePaiementId": {
            "type": "integer",
            "format": "int32"
          },
          "deviseId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SocieteDTO": {
        "required": [
          "adresse",
          "nomSociete",
          "numTelephone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nomSociete": {
            "type": "string",
            "nullable": true
          },
          "adresse": {
            "type": "string",
            "nullable": true
          },
          "numTelephone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VenteDTO": {
        "required": [
          "detailsPaiements"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dateDeVente": {
            "type": "string",
            "format": "date-time"
          },
          "prixDeVente": {
            "type": "number",
            "format": "double"
          },
          "quantiteVendu": {
            "type": "integer",
            "format": "int32"
          },
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "deviseId": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "detailsPaiements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetailsPaiementDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VenteDetailsDTO": {
        "type": "object",
        "properties": {
          "dateDeVente": {
            "type": "string",
            "format": "date-time"
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "marque": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "prixDeVente": {
            "type": "number",
            "format": "double"
          },
          "numBon": {
            "type": "string",
            "nullable": true
          },
          "typeArticle": {
            "type": "string",
            "nullable": true
          },
          "quantite": {
            "type": "integer",
            "format": "int32"
          },
          "typeDePaiement": {
            "type": "string",
            "nullable": true
          },
          "detailsPaiement": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}